CEDS-1687 Rename TransportDetailsView to BorderTransport
Plus rename of unit tests
