+++ test/views/declaration/BorderTransportViewSpec.scala
@@ -16,53 +16,38 @@
 
 package views.declaration
 
-import controllers.declaration.routes
-import controllers.util.SaveAndReturn
-import forms.declaration.DepartureTransport
-import forms.declaration.TransportCodes._
-import helpers.views.declaration.CommonMessages
+import base.Injector
+import forms.declaration.TransportDetails
 import models.Mode
 import play.api.data.Form
-import play.twirl.api.Html
+import play.api.i18n.MessagesApi
+import play.api.test.Helpers.stubMessages
+import services.cache.ExportsTestData
 import unit.tools.Stubs
 import views.declaration.spec.UnitViewSpec
-import views.html.components.fields.{field_radio, field_text}
-import views.html.declaration.departure_transport
+import views.html.declaration.border_transport
 import views.tags.ViewTest
 
 @ViewTest
-class BorderTransportViewSpec extends UnitViewSpec with CommonMessages with Stubs {
+class BorderTransportViewSpec extends UnitViewSpec with ExportsTestData with Stubs with Injector {
 
-  val form: Form[DepartureTransport] = DepartureTransport.form()
+  private val page = new border_transport(mainTemplate)
+  private val form: Form[TransportDetails] = TransportDetails.form()
 
-  private val borderTransportPage = new departure_transport(mainTemplate)
+  private def createView(mode: Mode = Mode.Normal, form: Form[TransportDetails] = form): Document =
+    page(mode, form)(journeyRequest(), stubMessages())
 
-  def createView(form: Form[DepartureTransport] = form): Html =
-    borderTransportPage(Mode.Normal, form)(request, messages)
-
-  "BorderTransport View" should {
+  "TransportDetails View" should {
     val view = createView()
 
-    "have defined translation for used labels" in {
-      val messages = realMessagesApi.preferred(request)
-      messages must haveTranslationFor("supplementary.transportInfo.title")
-      messages must haveTranslationFor("supplementary.transportInfo.title")
-      messages must haveTranslationFor(backCaption)
-      messages must haveTranslationFor(saveAndContinueCaption)
-      messages must haveTranslationFor(saveAndReturnCaption)
-      messages must haveTranslationFor("supplementary.transportInfo.borderTransportMode.header.hint")
-      messages must haveTranslationFor("supplementary.transportInfo.borderTransportMode.header")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.sea")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.road")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.rail")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.air")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.postalOrMail")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.fixedTransportInstallations")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.inlandWaterway")
-      messages must haveTranslationFor("supplementary.transportInfo.transportMode.unknown")
-
-      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.departure.header")
-      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.departure.header.hint")
+    "have proper messages for labels" in {
+      val messages = instanceOf[MessagesApi].preferred(journeyRequest())
+      messages must haveTranslationFor("supplementary.transportInfo.active.title")
+      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.crossingTheBorder.header")
+      messages must haveTranslationFor(
+        "supplementary.transportInfo.meansOfTransport.crossingTheBorder.nationality.header"
+      )
+      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.crossingTheBorder.header.hint")
       messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.IMOShipIDNumber")
       messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.nameOfVessel")
       messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.wagonNumber")
@@ -71,163 +56,140 @@ class BorderTransportViewSpec extends UnitViewSpec with CommonMessages with Stub
       messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.aircraftRegistrationNumber")
       messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.europeanVesselIDNumber")
       messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.nameOfInlandWaterwayVessel")
-
-      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.reference.header")
-      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.reference.hint")
+      messages must haveTranslationFor("supplementary.transportInfo.meansOfTransport.CrossingTheBorder.IDNumber.header")
+      messages must haveTranslationFor("supplementary.transportInfo.container")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.notPrePaid")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.other")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.accHolder")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.cash")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.creditCard")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.cheque")
+      messages must haveTranslationFor("standard.transportDetails.paymentMethod.eFunds")
     }
 
     "display page title" in {
-      view.getElementById("title").text() mustBe messages("supplementary.transportInfo.title")
+      view.getElementById("title").text() mustBe "supplementary.transportInfo.active.title"
     }
 
     "display 'Back' button that links to 'Warehouse' page" in {
       val backButton = view.getElementById("link-back")
-      backButton.text() mustBe messages(backCaption)
-      backButton must haveHref(routes.WarehouseIdentificationController.displayPage())
+
+      backButton.text() mustBe "site.back"
+      backButton.getElementById("link-back") must haveHref(
+        controllers.declaration.routes.DepartureTransportController.displayPage(Mode.Normal)
+      )
     }
 
     "display 'Save and continue' button on page" in {
-      view.getElementById("submit").text() mustBe messages(saveAndContinueCaption)
+      view.getElementById("submit").text() mustBe "site.save_and_continue"
     }
 
     "display 'Save and return' button on page" in {
-      val saveAndReturn = view.getElementById("submit_and_return")
-      saveAndReturn.text() mustBe messages(saveAndReturnCaption)
-      saveAndReturn must haveAttribute("name", SaveAndReturn.toString)
+      view.getElementById("submit_and_return").text() mustBe "site.save_and_come_back_later"
     }
 
-    "display 'Mode of Transport' section" which {
-
-      val section = view.getElementById("borderModeOfTransportCode")
-
-      "have correct hint" in {
-        section
-          .getElementsByClass("form-hint")
-          .text() mustBe messages("supplementary.transportInfo.borderTransportMode.header.hint")
+    "display 'Means of Transport' section" which {
+      "nationality picker" in {
+        view
+          .getElementById("meansOfTransportCrossingTheBorderNationality-label")
+          .text() mustBe "supplementary.transportInfo.meansOfTransport.crossingTheBorder.nationality.header"
       }
-
-      "have correct legend" in {
-        section
-          .getElementsByTag("legend")
-          .text() must startWith(messages("supplementary.transportInfo.borderTransportMode.header"))
+      "has label" in {
+        view
+          .getElementById("meansOfTransportCrossingTheBorderType-label")
+          .text() mustBe "supplementary.transportInfo.meansOfTransport.crossingTheBorder.header"
       }
-
-      "have 'Sea' option" in {
-        section.getElementById("Border_Sea-label").text() mustBe "supplementary.transportInfo.transportMode.sea"
+      "has hint" in {
+        view
+          .getElementById("meansOfTransportCrossingTheBorderType-hint")
+          .text() mustBe "supplementary.transportInfo.meansOfTransport.crossingTheBorder.header.hint"
       }
-
-      "have 'Road' option" in {
-        section.getElementById("Border_Road-label").text() mustBe "supplementary.transportInfo.transportMode.road"
-      }
-
-      "have 'Rail' option" in {
-        section.getElementById("Border_Rail-label").text() mustBe "supplementary.transportInfo.transportMode.rail"
-      }
-
-      "have 'Air' option" in {
-        section.getElementById("Border_Air-label").text() mustBe "supplementary.transportInfo.transportMode.air"
-      }
-
-      "have 'Postal or Mail' option" in {
-        section
-          .getElementById("Border_PostalOrMail-label")
-          .text() mustBe "supplementary.transportInfo.transportMode.postalOrMail"
-      }
-
-      "have 'Fixed transport installations' option" in {
-        section
-          .getElementById("Border_FixedTransportInstallations-label")
-          .text() mustBe "supplementary.transportInfo.transportMode.fixedTransportInstallations"
-      }
-
-      "have 'Inland waterway transport' option" in {
-        section
-          .getElementById("Border_InlandWaterway-label")
-          .text() mustBe "supplementary.transportInfo.transportMode.inlandWaterway"
-      }
-
-      "have 'Mode unknown' option" in {
-        section
-          .getElementById("Border_Unknown-label")
-          .text() mustBe "supplementary.transportInfo.transportMode.unknown"
-      }
-    }
-
-    "display 'Transport details type' section " which {
-
-      val section = view.getElementById("meansOfTransportOnDepartureType")
-
-      "have label" in {
-        section
-          .getElementById("meansOfTransportOnDepartureType-label")
-          .text() mustBe "supplementary.transportInfo.meansOfTransport.departure.header"
-      }
-
-      "have hint" in {
-        section
-          .getElementById("meansOfTransportOnDepartureType-hint")
-          .text() mustBe "supplementary.transportInfo.meansOfTransport.departure.header.hint"
-      }
-
-      "have 'Ship number' option" in {
-        section
-          .getElementById("Departure_IMOShipIDNumber-label")
+      "has 'Ship' section" in {
+        view
+          .getElementById("Border_IMOShipIDNumber-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.IMOShipIDNumber"
       }
-
-      "have 'Name of vessel' option" in {
-        section
-          .getElementById("Departure_NameOfVessel-label")
+      "has 'Vessel' section" in {
+        view
+          .getElementById("Border_NameOfVessel-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.nameOfVessel"
       }
-
-      "have 'Vagon number' option" in {
-        section
-          .getElementById("Departure_WagonNumber-label")
+      "has 'Vagon' section" in {
+        view
+          .getElementById("Border_WagonNumber-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.wagonNumber"
       }
-
-      "have 'Vehice number' option" in {
-        section
-          .getElementById("Departure_VehicleRegistrationNumber-label")
+      "has 'Register Vehicle' section" in {
+        view
+          .getElementById("Border_VehicleRegistrationNumber-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.vehicleRegistrationNumber"
       }
-
-      "have 'flight number' option" in {
-        section
-          .getElementById("Departure_IATAFlightNumber-label")
+      "has 'Fligh number' section" in {
+        view
+          .getElementById("Border_IATAFlightNumber-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.IATAFlightNumber"
       }
-
-      "have 'aircraft registration' option" in {
-        section
-          .getElementById("Departure_AircraftRegistrationNumber-label")
+      "has 'Aircraft Number' section" in {
+        view
+          .getElementById("Border_AircraftRegistrationNumber-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.aircraftRegistrationNumber"
       }
-
-      "have 'eni code' optopn" in {
-        section
-          .getElementById("Departure_EuropeanVesselIDNumber-label")
+      "has 'European Vessel' section" in {
+        view
+          .getElementById("Border_EuropeanVesselIDNumber-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.europeanVesselIDNumber"
       }
-
-      "have 'inland waterway' option" in {
-        section
-          .getElementById("Departure_NameOfInlandWaterwayVessel-label")
+      "has 'Inland waterway vessel' section " in {
+        view
+          .getElementById("Border_NameOfInlandWaterwayVessel-label")
           .text() mustBe "supplementary.transportInfo.meansOfTransport.nameOfInlandWaterwayVessel"
+      }
+      "has Reference input text" in {
+        view
+          .getElementById("meansOfTransportCrossingTheBorderIDNumber-label")
+          .text() mustBe "supplementary.transportInfo.meansOfTransport.CrossingTheBorder.IDNumber.header"
       }
     }
 
-    "display 'Reference' section" which {
-      "have label" in {
+    "display 'Container' section" in {
+      view.getElementById("container-label").text() mustBe "supplementary.transportInfo.container"
+    }
+
+    "display 'Payment Method' section" which {
+      "have 'Not PrePaid' option" in {
         view
-          .getElementById("meansOfTransportOnDepartureIDNumber-label")
-          .text() mustBe "supplementary.transportInfo.meansOfTransport.reference.header"
+          .getElementById("standard.transportDetails.paymentMethod.notPrePaid-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.notPrePaid"
       }
-      "have hint" in {
+      "have 'Other' option" in {
         view
-          .getElementById("meansOfTransportOnDepartureIDNumber-hint")
-          .text() mustBe "supplementary.transportInfo.meansOfTransport.reference.hint"
+          .getElementById("standard.transportDetails.paymentMethod.other-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.other"
+      }
+      "have 'Account Holder' option" in {
+        view
+          .getElementById("standard.transportDetails.paymentMethod.accHolder-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.accHolder"
+      }
+      "have 'Cash' option" in {
+        view
+          .getElementById("standard.transportDetails.paymentMethod.cash-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.cash"
+      }
+      "have 'Credit Card' option" in {
+        view
+          .getElementById("standard.transportDetails.paymentMethod.creditCard-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.creditCard"
+      }
+      "have 'Cheque' option" in {
+        view
+          .getElementById("standard.transportDetails.paymentMethod.cheque-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.cheque"
+      }
+      "have 'eFounds' option" in {
+        view
+          .getElementById("standard.transportDetails.paymentMethod.eFunds-label")
+          .text() mustBe "standard.transportDetails.paymentMethod.eFunds"
       }
     }
   }
