+++ test/views/declaration/TransportContainerSummaryViewSpec.scala
@@ -75,7 +75,7 @@ class TransportContainerSummaryViewSpec extends UnitViewSpec with Stubs with Mus
 
       backLinkContainer.text() must be(messages(backCaption))
       backLinkContainer.getElementById("link-back") must haveHref(
-        controllers.declaration.routes.TransportDetailsController.displayPage(Mode.Normal)
+        controllers.declaration.routes.BorderTransportController.displayPage(Mode.Normal)
       )
     }
